+++ kraken-std/src/kraken/std/python/tasks/update_pyproject_task.py
@@ -3,9 +3,9 @@ from __future__ import annotations
 from pathlib import Path
 
 from kraken.core import Project, Property, TaskStatus
-from kraken.core.lib.render_file_task import RenderFileTask
 
 from kraken.std.python.buildsystem import PythonBuildSystem
+from kraken.std.util.render_file_task import RenderFileTask
 
 from ..pyproject import Pyproject
 from ..settings import PythonSettings, python_settings
